refactor: move single operator parsing logic into Operator
