+++ ThePhasesOfCraftship/2_best_practice_first/testingBasics/exercises/3_Manage_Complexity/3_2_Boolean_Calculator/src/ExpressionParser.ts
@@ -5,9 +5,11 @@ import { FALSE, TRUE } from "./Value";
 
 const validWords = ['TRUE', 'FALSE', 'NOT', 'AND', 'OR'];
 
+export type PartiallyParsedExpression = (string | Evaluateable)[];
+
 export class ExpressionParser {
     static Parse(booleanString: string): Evaluateable {
-        const sections = splitByParentheses(booleanString).map((section, _, array) => {
+        const sections = splitByParentheses(booleanString).map((section) => {
             if(section[0] === '(') {
                 return this.Parse(section.substring(1, section.length - 1));
             }
@@ -22,41 +24,14 @@ export class ExpressionParser {
         return this.ParseWithoutParenthesis(sections);
     }
 
-    private static ParseOperator(expression: (string | Evaluateable)[], OperatorClass: new (...params:any) => Operator): (string | Evaluateable)[] {
-        const temp = [...expression];
-        const name = OperatorClass.name.toUpperCase();
-        const count = temp.filter(word => word === name).length;
-        const isUnary = OperatorClass.prototype instanceof UnaryOperator;
-        for (let i = 0; i < count; i++) {
-            const index = isUnary ? temp.lastIndexOf(name) : temp.indexOf(name);
-            if(isUnary) {
-                const value = temp[index + 1];
-                if(!(value instanceof Evaluateable)) {
-                    throw new Error('Invalid boolean string');
-                }
-                temp.splice(index, 2, new OperatorClass(value));
-            } else {
-                const left = temp[index - 1];
-                const right = temp[index + 1];
-                if(!(left instanceof Evaluateable && right instanceof Evaluateable)) {
-                    throw new Error('Invalid boolean string');
-                }
-
-                temp.splice(index - 1, 3, new OperatorClass(left, right));
-            }
-        }
-        
-        return temp;
-    }
-
-    private static ParseWithoutParenthesis(expression: (string | Evaluateable)[]): Evaluateable {
-        let temp: (string | Evaluateable)[] = expression.map(word=>{
+    private static ParseWithoutParenthesis(expression: PartiallyParsedExpression): Evaluateable {
+        let temp: PartiallyParsedExpression = expression.map(word=>{
             return word === 'TRUE' ? TRUE : word === 'FALSE' ? FALSE : word
         });
 
-        temp = this.ParseOperator(temp, Not);
-        temp = this.ParseOperator(temp, And);
-        temp = this.ParseOperator(temp, Or);
+        temp = Operator.Parse(temp, Not);
+        temp = Operator.Parse(temp, And);
+        temp = Operator.Parse(temp, Or);
 
         if(temp.length !== 1 || !(temp[0] instanceof Evaluateable)) {
             throw new Error('Invalid boolean string');
