+++ ThePhasesOfCraftship/2_best_practice_first/testingBasics/exercises/3_Manage_Complexity/3_2_Boolean_Calculator/src/Operator.ts
@@ -1,7 +1,17 @@
 import { Evaluateable } from "./Evaluateable";
+import { PartiallyParsedExpression } from "./ExpressionParser";
 
 export abstract class Operator extends Evaluateable {
+    static Parse(expression: PartiallyParsedExpression, OperatorClass: new (...params:any) => Operator): PartiallyParsedExpression{
+        return OperatorClass.prototype.parse(expression);
+    }
+
+    protected get name(): string {
+        return this.constructor.name.toUpperCase();
+    }
+
     abstract evaluate(): boolean;
+    protected abstract parse(expression: PartiallyParsedExpression): PartiallyParsedExpression;
 }
 
 export abstract class UnaryOperator extends Operator {
@@ -14,6 +24,21 @@ export abstract class UnaryOperator extends Operator {
     toString(): string {
         return `${this.constructor.name} ${this.value.toString()}`;
     }
+    protected parse(expression: PartiallyParsedExpression): PartiallyParsedExpression {
+        const temp = [...expression];
+        const count = temp.filter(word => word === this.name).length;
+       
+        for (let i = 0; i < count; i++) {
+            const index = temp.lastIndexOf(this.name);
+            const value = temp[index + 1];
+            if(!(value instanceof Evaluateable)) {
+                throw new Error('Invalid boolean string');
+            }
+            temp.splice(index, 2, new (this.constructor as any)(value));
+        }
+        
+        return temp;
+    }
 }
 
 abstract class BinaryOperator extends Operator {
@@ -25,9 +50,26 @@ abstract class BinaryOperator extends Operator {
         this.right = right;
     }
 
-
     toString(): string {
         return `${this.left.toString()} ${this.constructor.name} ${this.right.toString()}`;
+    }
+
+    protected parse(expression: PartiallyParsedExpression): PartiallyParsedExpression {
+        const temp = [...expression];
+        const count = temp.filter(word => word === this.name).length;
+       
+        for (let i = 0; i < count; i++) {
+            const index = temp.indexOf(this.name);
+            const left = temp[index - 1];
+            const right = temp[index + 1];
+            if(!(left instanceof Evaluateable && right instanceof Evaluateable)) {
+                throw new Error('Invalid boolean string');
+            }
+            
+            temp.splice(index - 1, 3, new (this.constructor as any)(left, right));
+        }
+        
+        return temp;
     }
 }
